+++ packages/core/src/utils/index.ts
@@ -1,5 +1,6 @@
 export { default as debounce } from './debounce';
 export { default as throttle } from './throttle';
+export type { Throttler } from './throttle.js';
 export { default as isMobileDevice } from './isMobileDevice';
 export { default as isTouchCapable } from './isTouchCapable';
 export { default as createFilter } from './createFilter';