+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cohensive/select-core
 
+## 0.14.3
+
+### Patch Changes
+
+- Update async options loading and abort controller behavior.
+
 ## 0.14.2
 
 ### Patch Changes